sstable/rowblk: move Iter
