+++ sstable/reader.go
@@ -406,7 +406,7 @@ func (r *Reader) NewRawRangeDelIter(transforms IterTransforms) (keyspan.Fragment
 	// sstables. This is because rangedels do not apply to points in the same
 	// sstable at the same sequence number anyway, so exposing obsolete rangedels
 	// is harmless.
-	if err := i.blockIter.initHandle(r.Compare, r.Split, h, transforms); err != nil {
+	if err := i.blockIter.InitHandle(r.Compare, r.Split, h, transforms); err != nil {
 		return nil, err
 	}
 	return keyspan.MaybeAssert(i, r.Compare), nil
@@ -433,7 +433,7 @@ func (r *Reader) NewRawRangeKeyIter(transforms IterTransforms) (keyspan.Fragment
 		return nil, err
 	}
 	i := newFragmentBlockIter(false /* elideSameSeqnum */)
-	if err := i.blockIter.initHandle(r.Compare, r.Split, h, transforms); err != nil {
+	if err := i.blockIter.InitHandle(r.Compare, r.Split, h, transforms); err != nil {
 		return nil, err
 	}
 	return keyspan.MaybeAssert(i, r.Compare), nil
@@ -627,8 +627,8 @@ func (r *Reader) transformRangeDelV1(b []byte) ([]byte, error) {
 	// v1 format range-del blocks have unfragmented and unsorted range
 	// tombstones. We need properly fragmented and sorted range tombstones in
 	// order to serve from them directly.
-	iter := &blockIter{}
-	if err := iter.init(r.Compare, r.Split, b, NoTransforms); err != nil {
+	iter := &rowblk.Iter{}
+	if err := iter.Init(r.Compare, r.Split, b, NoTransforms); err != nil {
 		return nil, err
 	}
 	var tombstones []keyspan.Span
@@ -808,7 +808,7 @@ func (r *Reader) Layout() (*Layout, error) {
 
 	if r.Properties.IndexPartitions == 0 {
 		l.Index = append(l.Index, r.indexBH)
-		iter, _ := newBlockIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
+		iter, _ := rowblk.NewIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
 		for kv := iter.First(); kv != nil; kv = iter.Next() {
 			dataBH, err := decodeBlockHandleWithProperties(kv.InPlaceValue())
 			if err != nil {
@@ -821,8 +821,8 @@ func (r *Reader) Layout() (*Layout, error) {
 		}
 	} else {
 		l.TopIndex = r.indexBH
-		topIter, _ := newBlockIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
-		iter := &blockIter{}
+		topIter, _ := rowblk.NewIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
+		iter := &rowblk.Iter{}
 		for kv := topIter.First(); kv != nil; kv = topIter.Next() {
 			indexBH, err := decodeBlockHandleWithProperties(kv.InPlaceValue())
 			if err != nil {
@@ -836,7 +836,7 @@ func (r *Reader) Layout() (*Layout, error) {
 				return nil, err
 			}
 			// TODO(msbutler): figure out how to pass virtualState to layout call.
-			if err := iter.init(r.Compare, r.Split, subIndex.Get(), NoTransforms); err != nil {
+			if err := iter.Init(r.Compare, r.Split, subIndex.Get(), NoTransforms); err != nil {
 				return nil, err
 			}
 			for kv := iter.First(); kv != nil; kv = iter.Next() {
@@ -850,7 +850,7 @@ func (r *Reader) Layout() (*Layout, error) {
 				l.Data = append(l.Data, dataBH)
 			}
 			subIndex.Release()
-			*iter = iter.resetForReuse()
+			*iter = iter.ResetForReuse()
 		}
 	}
 	if r.valueBIH.h.Length != 0 {
@@ -969,16 +969,16 @@ func (r *Reader) EstimateDiskUsage(start, end []byte) (uint64, error) {
 	// Iterators over the bottom-level index blocks containing start and end.
 	// These may be different in case of partitioned index but will both point
 	// to the same blockIter over the single index in the unpartitioned case.
-	var startIdxIter, endIdxIter *blockIter
+	var startIdxIter, endIdxIter *rowblk.Iter
 	if r.Properties.IndexPartitions == 0 {
-		iter, err := newBlockIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
+		iter, err := rowblk.NewIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
 		if err != nil {
 			return 0, err
 		}
 		startIdxIter = iter
 		endIdxIter = iter
 	} else {
-		topIter, err := newBlockIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
+		topIter, err := rowblk.NewIter(r.Compare, r.Split, indexH.Get(), NoTransforms)
 		if err != nil {
 			return 0, err
 		}
@@ -998,7 +998,7 @@ func (r *Reader) EstimateDiskUsage(start, end []byte) (uint64, error) {
 			return 0, err
 		}
 		defer startIdxBlock.Release()
-		startIdxIter, err = newBlockIter(r.Compare, r.Split, startIdxBlock.Get(), NoTransforms)
+		startIdxIter, err = rowblk.NewIter(r.Compare, r.Split, startIdxBlock.Get(), NoTransforms)
 		if err != nil {
 			return 0, err
 		}
@@ -1019,7 +1019,7 @@ func (r *Reader) EstimateDiskUsage(start, end []byte) (uint64, error) {
 				return 0, err
 			}
 			defer endIdxBlock.Release()
-			endIdxIter, err = newBlockIter(r.Compare, r.Split, endIdxBlock.Get(), NoTransforms)
+			endIdxIter, err = rowblk.NewIter(r.Compare, r.Split, endIdxBlock.Get(), NoTransforms)
 			if err != nil {
 				return 0, err
 			}
